デシベル計算を周波数別に分離して実施するように変更

+++ MicrophoneLevelLogger/Domain/IMasterPeakValues.cs
@@ -3,5 +3,5 @@
 public interface IMasterPeakValues
 {
     IMicrophone Microphone { get; }
-    IReadOnlyList<float> PeakValues { get; }
+    IReadOnlyList<double> PeakValues { get; }
 }
